Compensate an Opsview bug in fan perfdata
For iDracs with more then two fans, perf_data was included in the value itself from the third fan and on. That was generating an extra value for the same fan when its RPM changed, causing too many values (Fan IDs) and untraceable graph.
Now Fans output is consolidated to a single line after the "|" (pipe) separator so that Opsview keeps collecting the output as expected.